test.py
# ไฟล์: config/settings.py

import os
from dotenv import load_dotenv

load_dotenv() # โหลดค่าจากไฟล์ .env

# การเชื่อมต่อฐานข้อมูล
DATABASES = {
    'default': {
        'ENGINE': 'django.db.backends.postgresql',
        'NAME': os.getenv('DB_NAME', 'mood2movie'),
        'USER': os.getenv('DB_USER', 'postgres'),
        'PASSWORD': os.getenv('DB_PASSWORD'),
        'HOST': os.getenv('DB_HOST', 'localhost'),
        'PORT': os.getenv('DB_PORT', '5432'),
    }
}

# การกำหนดค่า API ภายนอก
TMDB_API_KEY = os.getenv('TMDB_API_KEY')
TMDB_BASE_URL = "https://api.themoviedb.org/3"